requirements.txt
pydantic~=2.10.6
openai~=1.66.3
tenacity~=9.0.0
pyyaml~=6.0.2
loguru~=0.7.3
numpy
datasets~=3.4.1
fastapi~=0.115.11
tiktoken~=0.9.0

html2text~=2024.2.26
gymnasium~=1.1.1
pillow~=11.1.0
browsergym~=0.13.3
uvicorn~=0.34.0
unidiff~=0.7.5
redis~=4.6.0
browser-use~=0.1.40
googlesearch-python~=1.3.0
baidusearch~=1.0.3
duckduckgo_search~=7.5.3

aiofiles~=24.1.0
pydantic_core~=2.27.2
colorama~=0.4.6
playwright~=1.51.0

docker~=7.1.0
pytest~=8.3.5
pytest-asyncio~=0.25.3

mcp~=1.5.0
httpx>=0.27.0
tomli>=2.0.0

boto3~=1.37.18

requests~=2.32.3
beautifulsoup4~=4.13.3

huggingface-hub~=0.29.2
setuptools~=75.8.0

# Database and ORM
sqlalchemy~=2.0.25
alembic~=1.13.1
aiosqlite~=0.19.0
asyncpg~=0.29.0

# Authentication and Security
python-jose[cryptography]~=3.3.0
passlib[bcrypt]~=1.7.4

# WebSocket support
websockets~=12.0

# Redis for caching
# Redis with Python 3.12+ compatibility
redis>=4.6.0

# Additional utilities
python-dotenv~=1.0.0
